fix #23: Segfault when binding CLOB/BLOB to PL/SQL OUT parameters
This is a bug in the internal method OCI_BindCheck() of the C API.
In case of OCI_BDM_OUT only bind direction mode, the method does not set
the content of the loblocator handle pointer passed to Oracle client.
Then the Oracle client is trying to set the content of that pointer to
the value returned from the procedure. As the pointer is NULL, a
segfault occurs !